+++ packages/app/src/app.vue
@@ -756,6 +756,14 @@ icon
     background-image: url("./images/open.svg");
 }
 
+.mcr-icon-link {
+    padding-left: 18px;
+    background-image: url("./images/link.svg");
+    background-repeat: no-repeat;
+    background-position: left center;
+    background-size: 16px 16px;
+}
+
 .mcr-header {
     color: #fff;
     background-color: #24292f;